Moved Spherical and Polar to math.polar.ts
